build: polish tsup
