+++ package.json
@@ -34,7 +34,7 @@
     "clean": "pnpm -r --filter=./packages/** run clean",
     "format": "prettier --check \"packages/*/src/**/*.{ts,js,vue}\"",
     "lint": "eslint \"packages/*/src/**/*.{ts,js}\"",
-    "publint": "pnpm -r --filter='vite-plugin-checker' exec publint",
+    "publint": "pnpm -r --filter './packages/**' --filter=!@vite-plugin-checker/runtime exec publint",
     "test": "run-s test-unit test-serve test-build",
     "test-unit": "vitest run",
     "test-serve": "vitest run -c vitest.config.e2e.ts",
@@ -59,49 +59,50 @@
     ]
   },
   "devDependencies": {
-    "@changesets/cli": "^2.26.1",
-    "@tsconfig/esm": "^1.0.2",
-    "@tsconfig/node18": "^1.0.1",
-    "@tsconfig/strictest": "^2.0.0",
-    "@types/babel__code-frame": "^7.0.2",
-    "@types/debug": "^4.1.5",
-    "@types/fs-extra": "^11.0.1",
-    "@types/klaw": "^3.0.3",
-    "@types/minimist": "^1.2.2",
+    "@changesets/cli": "^2.27.1",
+    "@tsconfig/esm": "^1.0.5",
+    "@tsconfig/node18": "^18.2.2",
+    "@tsconfig/strictest": "^2.0.2",
+    "@types/babel__code-frame": "^7.0.6",
+    "@types/debug": "^4.1.12",
+    "@types/fs-extra": "^11.0.4",
+    "@types/klaw": "^3.0.6",
+    "@types/minimist": "^1.2.5",
     "@types/node": "^18.0.0",
-    "@types/prompts": "^2.0.13",
+    "@types/prompts": "^2.4.9",
     "@types/rimraf": "^3.0.0",
-    "@types/semver": "^7.3.6",
-    "@types/ws": "^8.5.3",
-    "@typescript-eslint/eslint-plugin": "^5.59.0",
-    "@typescript-eslint/parser": "^5.59.0",
+    "@types/semver": "^7.5.6",
+    "@types/ws": "^8.5.10",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
     "chalk": "^4.1.1",
     "cross-env": "^7.0.3",
-    "eslint": "^8.51.0",
-    "eslint-config-alloy": "^4.5.1",
+    "eslint": "^8.56.0",
+    "eslint-config-alloy": "^5.1.2",
     "execa": "^5.1.1",
-    "fast-glob": "^3.2.7",
+    "fast-glob": "^3.3.2",
     "fast-json-stable-stringify": "^2.1.0",
-    "fs-extra": "^11.1.0",
+    "fs-extra": "^11.2.0",
     "jest-serializer-path": "^0.1.15",
-    "klaw": "^4.0.1",
+    "klaw": "^4.1.0",
     "lint-staged": "^11.0.0",
-    "minimist": "^1.2.5",
+    "minimist": "^1.2.8",
     "npm-run-all": "^4.1.5",
-    "playwright-chromium": "^1.24.0",
+    "playwright-chromium": "^1.41.1",
     "prettier": "^2.3.2",
     "prompts": "^2.4.1",
-    "publint": "^0.1.9",
+    "publint": "^0.2.7",
     "rimraf": "^3.0.2",
-    "semver": "^7.5.0",
-    "simple-git-hooks": "^2.8.0",
+    "semver": "^7.5.4",
+    "simple-git-hooks": "^2.9.0",
     "sort-deep-object-arrays": "^1.1.2",
-    "strip-ansi": "^7.0.0",
+    "strip-ansi": "^7.1.0",
+    "tsup": "^6.7.0",
     "tiny-invariant": "^1.3.1",
     "typescript": "^5.0.4",
     "vite": "^5.0.12",
     "vite-plugin-checker": "workspace:*",
-    "vitest": "^0.30.1",
-    "ws": "^8.5.0"
+    "vitest": "^1.2.2",
+    "ws": "^8.16.0"
   }
 }
