+++ packages/checker-eslint/package.json
@@ -4,13 +4,13 @@
   "private": true,
   "description": "ESLint checker for vite-plugin-checker",
   "types": "./dist/esm/main.d.ts",
-  "main": "./dist/cjs/main.js",
+  "main": "./dist/cjs/main.cjs",
   "type": "module",
   "exports": {
     ".": {
       "types": "./dist/esm/main.d.ts",
-      "import": "./dist/esm/main.js",
-      "require": "./dist/cjs/main.js"
+      "import": "./dist/esm/main.mjs",
+      "default": "./dist/cjs/main.cjs"
     }
   },
   "engines": {
@@ -48,7 +48,6 @@
   "devDependencies": {
     "@types/eslint": "^8.44.0",
     "eslint": "^8.51.0",
-    "tsup": "^6.7.0",
     "vite-plugin-checker": "workspace:*"
   }
 }
