+++ packages/checker-vls/package.json
@@ -4,13 +4,13 @@
   "private": true,
   "description": "TypeScript checker for vite-plugin-checker",
   "types": "./dist/esm/main.d.ts",
-  "main": "./dist/cjs/main.js",
+  "main": "./dist/cjs/main.cjs",
   "type": "module",
   "exports": {
     ".": {
       "types": "./dist/esm/main.d.ts",
-      "import": "./dist/esm/main.js",
-      "require": "./dist/cjs/main.js"
+      "import": "./dist/esm/main.mjs",
+      "default": "./dist/cjs/main.cjs"
     }
   },
   "engines": {
@@ -51,7 +51,6 @@
   "devDependencies": {
     "vite-plugin-checker": "workspace:*",
     "vls": "^0.8.1",
-    "tsup": "^6.7.0",
     "typescript": "^5.0.4"
   }
 }
