+++ packages/runtime/package.json
@@ -18,6 +18,6 @@
     "@vitejs/plugin-vue": "^5.0.3",
     "vite": "^5.0.12",
     "vue": "^3.3.4",
-    "vue-tsc": "^1.6.1"
+    "vue-tsc": "^1.8.27"
   }
 }
