+++ playground/vue-tsc-vue3/package.json
@@ -17,6 +17,6 @@
     "vite": "^5.0.12",
     "vite-plugin-checker": "workspace:*",
     "@vite-plugin-checker/vue-tsc": "workspace:*",
-    "vue-tsc": "^1.6.1"
+    "vue-tsc": "^1.8.27"
   }
 }
